refactor: add business rules on event.service

+++ backend/src/event/event.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  NotFoundException,
+  Injectable,
+} from '@nestjs/common';
 
 import { CreateEventDto } from './dto/create-event-dto';
 import { UpdateEventDto } from './dto/update-event-dto';
@@ -12,23 +16,86 @@ export class EventService {
     return 'Event Service is healthy';
   }
 
-  create(createEventDto: CreateEventDto) {
-    return this.eventRepository.create(createEventDto);
+  async create(createEventDto: CreateEventDto) {
+    if (!createEventDto.name) {
+      throw new BadRequestException('Nome é obrigatório');
+    }
+    if (!createEventDto.date) {
+      throw new BadRequestException('Data é obrigatória');
+    }
+    if (!createEventDto.location) {
+      throw new BadRequestException('Local é obrigatório');
+    }
+    const eventDate = new Date(createEventDto.date);
+
+    if (eventDate < new Date()) {
+      throw new BadRequestException('Data não pode ser no passado');
+    }
+
+    return await this.eventRepository.create(createEventDto);
   }
 
-  findAll() {
-    return this.eventRepository.findAll();
+  async findAll() {
+    return await this.eventRepository.findAll();
   }
 
-  findOne(id: number) {
-    return this.eventRepository.findOne(id);
+  async findOne(id: number) {
+    if (!id) {
+      throw new BadRequestException('ID é obrigatório');
+    }
+    if (isNaN(id)) {
+      throw new BadRequestException('ID deve ser um número');
+    }
+    const event = await this.eventRepository.findOne(id);
+    if (!event) {
+      throw new NotFoundException(`Evento com ID ${id} não encontrado`);
+    }
+
+    return event;
   }
 
-  update(id: number, updateEventDto: UpdateEventDto) {
-    return this.eventRepository.update(id, updateEventDto);
+  async update(id: number, updateEventDto: UpdateEventDto) {
+    if (!id) {
+      throw new BadRequestException('ID é obrigatório');
+    }
+    if (isNaN(id)) {
+      throw new BadRequestException('ID deve ser um número');
+    }
+    if (!updateEventDto.name) {
+      throw new BadRequestException('Nome é obrigatório');
+    }
+    if (!updateEventDto.date) {
+      throw new BadRequestException('Data é obrigatória');
+    }
+    if (!updateEventDto.location) {
+      throw new BadRequestException('Local é obrigatório');
+    }
+
+    const event = await this.eventRepository.findOne(id);
+    if (!event) {
+      throw new NotFoundException(`Evento com ID ${id} não encontrado`);
+    }
+    const eventDate = new Date(updateEventDto.date);
+
+    if (eventDate < new Date()) {
+      throw new BadRequestException('Data não pode ser no passado');
+    }
+
+    return await this.eventRepository.update(id, updateEventDto);
   }
 
-  remove(id: number) {
-    return this.eventRepository.remove(id);
+  async remove(id: number) {
+    if (!id) {
+      throw new BadRequestException('ID é obrigatório');
+    }
+    if (isNaN(id)) {
+      throw new BadRequestException('ID deve ser um número');
+    }
+    const event = await this.eventRepository.findOne(id);
+    if (!event) {
+      throw new NotFoundException(`Evento com ID ${id} não encontrado`);
+    }
+
+    return await this.eventRepository.remove(id);
   }
 }
